% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cdpathC}
\alias{cdpathC}
\title{compute solutions to the penalized likelihood estimation
problem on a grid of lambdas}
\usage{
cdpathC(node, data, obslist, eorder, fmlam = 0.001, nlam = 50,
  eps = 1e-04, beta, lambda, gamma = 0.15, lowerbound = 1e-04)
}
\arguments{
\item{node}{the number of nodes of the DAG}

\item{data}{the data matrix of dimension (nobs*noi)*node. See DatGen}

\item{obslist}{list of observational samples. See DatGen}

\item{eorder}{matrix indicating the order to cycle through
the pair of nodes in the CD algorithm. nrow of the matrix is
0.5*node*(node-1), and ncol of the matrix is 2.
Each row is a pair of nodes}

\item{fmlam}{the ratio of lambda_min/lambda_max}

\item{nlam}{the number of lambda values}

\item{eps}{threshold of convergence for the CD algorithm}

\item{beta}{the initial coefficient matrix B^0.
If missing, will assume to be all zero matrix}

\item{lambda}{the sequence of lambda values to be used.
If missing, it will be constructed from data using fmlam and nlam.}

\item{gamma}{the exponent in the weights for the adaptive lasso penalty}

\item{lowerbound}{the constant used to truncate the weights
for the adaptive lasso penalty}
}
\value{
A list of 2
\item{coef}{matrix of estimated coefficients, of dimension nlam*(node^2).
            Each row contains the estimate of the coefficient
            matrix (stored as a vector) at a single value of lambda}
\item{lambda}{the sequence of lambda values used}
}
\description{
compute solutions to the penalized likelihood estimation
problem on a grid of lambdas
}

