% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DatGen}
\alias{DatGen}
\title{Generate Gaussian interventional data from DAGs}
\usage{
DatGen(ordex, ts, coeff, noi, nobs)
}
\arguments{
\item{ordex}{parent set matrix, of dimension maxdeg*node. See GGen}

\item{ts}{one topological sort of the DAG. See GGen}

\item{coeff}{the beta value used to generate Gaussian samplesb$ob}

\item{noi}{the number of nodes that have interventional smaples
(i.e., the number of data blocks). Each data block contains
interventional samples where one particular node is experimentally manipulated}

\item{nobs}{the sample size of each data block}
}
\value{
A list of 3
\item{data}{data matrix of dimension (nobs*noi)*node}
\item{ivn}{the vector of nodes that have interventional samples}
\item{obslist}{list of indices of observational samples for each node
               (excluding interventional samples).
               The list has length equal to node}
}
\description{
Generate Gaussian interventional data from DAGs
}

