% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GGen}
\alias{GGen}
\title{Generate DAGs given the maximum in-degree and the total number of edges}
\usage{
GGen(maxdeg, node, nedge)
}
\arguments{
\item{maxdeg}{maximum in-degree}

\item{node}{the number of nodes}

\item{nedge}{the total number of edges}
}
\value{
A list of 4.
\item{ordex}{parent set matrix, of dimension maxdeg*node.
             Each nonzero entry in column j lists one parent of node j.
             The parents	of node j are sorted in ascending order.}
\item{trueG}{the adjacency matrix of the generated DAG}
\item{ts}{one topological sort of the generated DAG}
\item{degree.dist}{the distribution of in-degrees of all nodes.
                   Element j lists the number of nodes whose in-degree equals j-1}
}
\description{
Generate DAGs given the maximum in-degree and the total number of edges
}

