% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plog}
\alias{plog}
\title{compute the negative log-likelihood using unpenalized coefficients}
\usage{
plog(node, G, data, obslist)
}
\arguments{
\item{node}{the number of nodes of the DAG}

\item{G}{the adjacency matrix of the DAG}

\item{data}{the data matrix of dimension (nobs*noi)*node. See DatGen}

\item{obslist}{list of observational samples. See DatGen}
}
\value{
pLog, the negative log-likelihood
}
\description{
compute the negative log-likelihood using unpenalized coefficients
}

